feat(db): run the cache and tracker tables on Neon, from a Node function
The site is live on wenjian.yishuziyu.cn but every shared table was a no-op, so
/api/s/entire always came back empty and each card fetched its own source. Four
separate things had to be true for the database to work here, and all four were
wrong:

- the Vercel branch of the nitro config forced the vercel-edge preset. Edge has no
  node:net, so a Postgres driver cannot run at all; it now uses the Node.js preset.
- nitro's own `database` option imports the connector by name at runtime, and the
  Vercel function has no resolvable node_modules at that point. The import failed,
  getCacheTable() swallowed it, and the cache silently never engaged. The database
  is now created from a connector that is imported statically.
- the local SQLite driver was better-sqlite3, a native module: the function entry
  statically imported it, loaded a macOS binary on Linux, and died before serving
  anything. Local dev uses node:sqlite now, loaded lazily.
- the shared tables were built for SQLite: INTEGER cannot hold millisecond
  timestamps in Postgres, `user` is a reserved word, INSERT OR REPLACE does not
  exist, and `?` placeholders have to become `$1..$n`. All four are handled, and
  the connector for Neon talks HTTP instead of holding a socket that the pooler
  can reap out from under an unhandled error event.

Also dropped the catch-all ISR rule: wrapping /api/s in an ISR function made it
fail every time (the same artifact answers 200 when run directly).

DEVLOG.md records which product owns which database, and docs/handoff-2026-09-12.md
passes the whole thing on.

+++ server/database/tracker.ts
@@ -15,13 +15,13 @@ export class TrackerTable {
         id TEXT PRIMARY KEY,
         user_id TEXT,
         topic TEXT,
-        days INTEGER,
-        interval_ms INTEGER,
-        enabled INTEGER,
-        created INTEGER,
-        updated INTEGER,
-        last_run INTEGER,
-        next_run INTEGER
+        days BIGINT,
+        interval_ms BIGINT,
+        enabled BIGINT,
+        created BIGINT,
+        updated BIGINT,
+        last_run BIGINT,
+        next_run BIGINT
       );
     `).run()
     await this.db.prepare(`
@@ -32,10 +32,10 @@ export class TrackerTable {
         topic TEXT,
         summary TEXT,
         model TEXT,
-        mock INTEGER,
-        source_count INTEGER,
+        mock BIGINT,
+        source_count BIGINT,
         steps TEXT,
-        created INTEGER
+        created BIGINT
       );
     `).run()
     logger.success(`init trackers + briefings tables`)
